如果数据采集，实体没有变化，停止实体更新

+++ src/main/java/com/snowcattle/game/db/service/entity/IEntityService.java
@@ -10,7 +10,7 @@ import java.util.List;
  * Created by jwp on 2017/3/23.
  */
 public interface IEntityService <T extends BaseEntity>{
-    public int insertEntity(T entity);
+    public long insertEntity(T entity);
     public IEntity getEntity(T entity);
     public List<T> getEntityList(T entity);
     public void updateEntity(T entity);
