+++ src/test/java/com/snowcattle/game/db/service/jdbc/service/IOrderService.java
@@ -8,7 +8,7 @@ import java.util.List;
  * Created by jiangwenping on 17/3/20.
  */
 public interface IOrderService {
-    public int insertOrder(Order order);
+    public long insertOrder(Order order);
     public Order getOrder(long userId, long id);
     public List<Order> getOrderList(long userId);
     void updateOrder(Order order);
